+++ lib/async_notifier.dart
@@ -1,0 +1,1 @@
+export 'package:async_notifier/async_notifier.dart';