+++ lib/src/extensions/element.dart
@@ -4,23 +4,22 @@ import 'package:flutter/widgets.dart';
 extension ContextElementExtension on BuildContext {
   /// Visits the [Element] tree below this [BuildContext].
   ///
-  /// If [onElement], [onState] or [onWidget] returns true, the visit will stop.
+  /// If [onElement], [onState] or [onWidget] returns `false`, the visit will stop.
   void visit({
     bool Function(Element element)? onElement,
     bool Function(Widget widget)? onWidget,
     bool Function(State state)? onState,
   }) {
-    assert(
-      onElement != null || onState != null || onWidget != null,
-      'At least one visitor must be set.',
-    );
+    if (onElement == null && onState == null && onWidget == null) {
+      return;
+    }
 
-    void visit(Element el) {
-      if (onElement?.call(el) ?? false) return;
-      if (onWidget?.call(el.widget) ?? false) return;
-      if (el is StatefulElement && (onState?.call(el.state) ?? false)) return;
+    void visit(Element e) {
+      if (onElement?.call(e) == false) return;
+      if (onWidget?.call(e.widget) == false) return;
+      if (e is StatefulElement && (onState?.call(e.state) == false)) return;
 
-      el.visitChildren(visit);
+      e.visitChildren(visit);
     }
 
     visitChildElements(visit);