+++ lib/src/extensions/future.dart
@@ -49,7 +49,7 @@ typedef SnackBarBuilder =
 extension AsyncSnackBar<T> on Future<T> {
   /// The default success snackbar to use on [showSnackBar].
   static SnackBarBuilder successBuilder = (context, message) {
-    return SnackBar(content: Text(message));
+    return SnackBar(content: SelectableText(message));
   };
 
   /// The default error snackbar to use on [showSnackBar].
@@ -57,13 +57,31 @@ extension AsyncSnackBar<T> on Future<T> {
     final theme = Theme.of(context);
 
     return SnackBar(
-      content: Text(message),
+      content: SelectableText(message),
       backgroundColor: theme.colorScheme.error,
     );
   };
 
   /// The default [AnimationStyle] to use on [showSnackBar].
   static AnimationStyle? animationStyle;
+
+  /// Shows a success [SnackBar].
+  /// You can customize the snackbar by setting [AsyncSnackBar.successBuilder].
+  static void showSuccess(String message, {BuildContext? context}) {
+    ScaffoldMessenger.of(context ??= Async.context).showSnackBar(
+      successBuilder(context, message),
+      snackBarAnimationStyle: animationStyle,
+    );
+  }
+
+  /// Shows an error [SnackBar].
+  /// You can customize the snackbar by setting [AsyncSnackBar.errorBuilder].
+  static void showError(Object error, {BuildContext? context}) {
+    ScaffoldMessenger.of(context ??= Async.context).showSnackBar(
+      errorBuilder(context, Async.message(error)),
+      snackBarAnimationStyle: animationStyle,
+    );
+  }
 
   /// Shows a [SnackBar] on success or error.
   ///
@@ -81,26 +99,92 @@ extension AsyncSnackBar<T> on Future<T> {
     String? successMessage,
     String? errorMessage,
   }) {
-    context ??= Async.context;
-    final messenger = ScaffoldMessenger.of(context);
-
     return then(
       (value) {
         if (successMessage != null) {
-          messenger.showSnackBar(
-            successBuilder(context!, successMessage),
-            snackBarAnimationStyle: animationStyle,
-          );
+          showSuccess(successMessage, context: context);
         }
         return value;
       },
       onError: (Object e, StackTrace s) {
-        messenger.showSnackBar(
-          errorBuilder(context!, errorMessage ?? Async.message(e)),
-          snackBarAnimationStyle: animationStyle,
-        );
+        showError(errorMessage ?? e, context: context);
         Error.throwWithStackTrace(e, s);
       },
     );
+  }
+}
+
+final _debouncers = <Object, (Completer<dynamic>, Timer)>{};
+final _throttlers = <Object, (Completer<dynamic>, Timer)>{};
+
+extension FutureFunctionExtensions<T, R> on Future<R> Function() {
+  Future<R> Function() debounceTime(Duration duration) {
+    return () {
+      final debouncer = _debouncers[this]?..$2.cancel();
+
+      final cp = debouncer?.$1 as Completer<R>? ?? Completer();
+      _debouncers[this] = (cp, Timer(duration, () => cp.complete(this())));
+
+      return cp.future.whenComplete(() => _debouncers.remove(this));
+    };
+  }
+
+  Future<R> Function() throttleTime(Duration duration) {
+    return () {
+      if (_throttlers[this]?.$1.future case Future<R> future) return future;
+
+      final cp = Completer<R>()..complete(this());
+      _throttlers[this] = (cp, Timer(duration, () => _throttlers.remove(this)));
+
+      return cp.future;
+    };
+  }
+}
+
+extension FutureFunctionExtensions1<T, R> on Future<R> Function(T) {
+  Future<R> Function(T) debounceTime(Duration duration) {
+    return (T arg) {
+      final debouncer = _debouncers[this]?..$2.cancel();
+
+      final cp = debouncer?.$1 as Completer<R>? ?? Completer();
+      _debouncers[this] = (cp, Timer(duration, () => cp.complete(this(arg))));
+
+      return cp.future.whenComplete(() => _debouncers.remove(this));
+    };
+  }
+
+  Future<R> Function(T) throttleTime(Duration duration) {
+    return (T arg) {
+      if (_throttlers[this]?.$1.future case Future<R> future) return future;
+
+      final cp = Completer<R>()..complete(this(arg));
+      _throttlers[this] = (cp, Timer(duration, () => _throttlers.remove(this)));
+
+      return cp.future;
+    };
+  }
+}
+
+extension FutureFunctionExtensions2<T1, T2, R> on Future<R> Function(T1, T2) {
+  Future<R> Function(T1, T2) debounceTime(Duration duration) {
+    return (T1 a, T2 b) {
+      final debouncer = _debouncers[this]?..$2.cancel();
+
+      final cp = debouncer?.$1 as Completer<R>? ?? Completer();
+      _debouncers[this] = (cp, Timer(duration, () => cp.complete(this(a, b))));
+
+      return cp.future.whenComplete(() => _debouncers.remove(this));
+    };
+  }
+
+  Future<R> Function(T1, T2) throttleTime(Duration duration) {
+    return (T1 a, T2 b) {
+      if (_throttlers[this]?.$1.future case Future<R> future) return future;
+
+      final cp = Completer<R>()..complete(this(a, b));
+      _throttlers[this] = (cp, Timer(duration, () => _throttlers.remove(this)));
+
+      return cp.future;
+    };
   }
 }